refactor: made textfield stateless

+++ lib/screens/enter_screen/widgets/buttons/delete_button.dart
@@ -8,6 +8,15 @@ class EnterScreenDeleteButton extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Consumer<EnterScreenViewModel>(builder: (context, viewModel, _) {
+      if (viewModel.initialTransaction == null && viewModel.initialSerialTransaction == null) {
+        return Container(
+          padding: const EdgeInsets.all(15),
+          child: const Icon(
+            Icons.delete_outline_rounded,
+            color: Colors.white,
+          ),
+        );
+      }
       return GestureDetector(
         onTap: () => viewModel.delete(),
         child: Container(
